requirements.txt
asgiref==3.7.2
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==23.3.0
iniconfig==2.0.0
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.10.0
